reorganize summary module
